Adressing code review comments
